Hello,
Is it possible to use binary search for `func (a *alphabet) Index(t rune) (int64, error)`?

Thanks