@jerome-laforge 2024-02-28
You are right.
Even with biggest random shortuuid list, I have slower `decoding`.
So we can reject this PR.